Add AbstractAsyncHistoryManager with common methods for adding fields to data and extract common methods for getting process definition into the AbstractHistoryManager
